Fix `getTimezoneOffset()` when `tm_gmtoff` is not available
In my environment, the build fails because `tm_gmtoff` is not available:

```
quickjs.c: In function 'getTimezoneOffset':
quickjs.c:40527:15: error: 'struct tm' has no member named 'tm_gmtoff'
40527 |     return -tm.tm_gmtoff / 60;
      |               ^
```

This approach attempts to reimplement the `getTimezoneOffset()` to not
rely on that property. This change works for my build environment, but
I am open to any suggestions for improvement.

I didn't test, but this might also work on Windows where this logic is
currently not implemented.